Moved dataGen, cassandra and postgres to folders
